+++ src/model/shape/Ellipse.java
@@ -7,7 +7,7 @@ import model.utils.Time;
 public class Ellipse extends AShape {
 
   public Ellipse(String name, model.shape.Shape type, Color col, double posX, double posY,
-                 int width, int height, Time time) {
+                 double width, double height, Time time) {
     super(name, type, col, posX, posY, width, height, time);
 
   }
